add configurable log level to api
