+++ helm/kre/templates/engine/admin-api/deployment.yaml
@@ -52,6 +52,8 @@ spec:
               valueFrom:
                 fieldRef:
                   fieldPath: metadata.namespace
+            - name: KRE_API_LOG_LEVEL
+              value: {{ .Values.adminApi.logLevel }}
           ports:
             - containerPort: 8080
               protocol: TCP
