+++ helm/kre/values.yaml
@@ -77,6 +77,7 @@ config:
     cookieDomain: kre.local
 
 adminApi:
+  logLevel: INFO
   image:
     repository: konstellation/kre-admin-api
     tag: latest
